refactor: rename threeStatsService -> threeStats.service
ref #2318
