+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeViewer.module.ts
@@ -2,7 +2,7 @@ import { ThreeViewerService } from "./threeViewerService"
 import { ThreeSceneService } from "./threeSceneService"
 import { ThreeCameraService } from "./threeCameraService"
 import { ThreeOrbitControlsService } from "./threeOrbitControlsService"
-import { ThreeStatsService } from "./threeStatsService"
+import { ThreeStatsService } from "./threeStats.service"
 import "../../../state/state.module"
 import camelCase from "lodash.camelcase"
